ARM: OMAP2+: powerdomain: skip register reads for powerdomains known to be on
There's no need to determine the current power state for powerdomains
that must be on while the kernel is running.  We mark these
powerdomains with a new flag, PWRDM_ACTIVE_WITH_KERNEL.  Any
powerdomain marked with that flag is reported as being in the ON power
state while the kernel is running.

The original version of this patch incorrectly listed the CORE powerdomain
as being active with the kernel on OMAP44xx.  This has been corrected
thanks to Jon Hunter <jon-hunter@ti.com>.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Cc: Benoît Cousson <b-cousson@ti.com>
Cc: Jon Hunter <jon-hunter@ti.com>