ARM: OMAP5: hwmod-data: Add timer clock activity flags
Currently HWMOD is not configuring the clock-activity field in the
OCP_CFG register for timers that have this field. It is needed to
keep the wake-up capability enabled.

Hence add the appropriate flags to the timer HWMOD structures, same way
as the commit 10759e82 {ARM: OMAP2+: Don't use __omap_dm_timer_reset()}.
On OMAP5 based SOCs, dmtimer 1, 2 and 10 have the clock-activity field.

Cc: Jon Hunter <jon-hunter@ti.com>
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>